app/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-core
langchain-google-vertexai
chroma
pypdf2